+++ docs/examples/ex_secret_types.py
@@ -6,14 +6,18 @@ class SimpleModel(BaseModel):
     password: SecretStr
     password_bytes: SecretBytes
 
-print(SimpleModel(password='IAmSensitive', password_bytes=b'IAmSensitiveBytes'))
+sm = SimpleModel(password='IAmSensitive', password_bytes=b'IAmSensitiveBytes')
+print(sm)
 # > SimpleModel password=SecretStr('**********') password_bytes=SecretBytes(b'**********')
 
-sm = SimpleModel(password='IAmSensitive', password_bytes=b'IAmSensitiveBytes')
 print(sm.password.get_secret_value())
 # > IAmSensitive
 print(sm.password_bytes.get_secret_value())
 # > b'IAmSensitiveBytes'
+print(sm.password.display())
+# > '**********'
+print(sm.json())
+# > '{"password": "**********", "password_bytes": "**********"}'
 
 
 try: