+++ pydantic/json.py
@@ -2,9 +2,12 @@ import datetime
 from decimal import Decimal
 from enum import Enum
 from ipaddress import IPv4Address, IPv4Interface, IPv4Network, IPv6Address, IPv6Interface, IPv6Network
+from pathlib import Path
 from types import GeneratorType
 from typing import Any, Callable, Dict, Type, Union
 from uuid import UUID
+
+from pydantic.types import SecretBytes, SecretStr
 
 __all__ = 'pydantic_encoder', 'custom_pydantic_encoder', 'timedelta_isoformat'
 
@@ -20,6 +23,8 @@ ENCODERS_BY_TYPE: Dict[Type[Any], Callable[[Any], Any]] = {
     IPv6Interface: str,
     IPv4Network: str,
     IPv6Network: str,
+    SecretStr: lambda ss: ss.display(),
+    SecretBytes: lambda sb: sb.display(),
     UUID: str,
     datetime.datetime: isoformat,
     datetime.date: isoformat,
@@ -40,6 +45,8 @@ def pydantic_encoder(obj: Any) -> Any:
         return obj.dict()
     elif isinstance(obj, Enum):
         return obj.value
+    elif isinstance(obj, Path):
+        return str(obj)
 
     try:
         encoder = ENCODERS_BY_TYPE[type(obj)]