+++ pydantic/types.py
@@ -89,9 +89,12 @@ NetworkType = Union[str, bytes, int, Tuple[Union[str, bytes, int], Union[str, in
 
 
 if TYPE_CHECKING:  # pragma: no cover
+    from .dataclasses import DataclassType  # noqa: F401
+    from .main import BaseModel  # noqa: F401
     from .utils import AnyCallable
 
     CallableGenerator = Generator[AnyCallable, None, None]
+    ModelOrDc = Type[Union['BaseModel', 'DataclassType']]
 
 
 class StrictStr(str):
@@ -592,6 +595,9 @@ class SecretStr:
     def __str__(self) -> str:
         return repr(self)
 
+    def display(self) -> str:
+        return '**********' if self._secret_value else ''
+
     def get_secret_value(self) -> str:
         return self._secret_value
 
@@ -614,6 +620,9 @@ class SecretBytes:
 
     def __str__(self) -> str:
         return repr(self)
+
+    def display(self) -> str:
+        return '**********' if self._secret_value else ''
 
     def get_secret_value(self) -> bytes:
         return self._secret_value