+++ tests/test_json.py
@@ -3,12 +3,14 @@ import json
 from decimal import Decimal
 from enum import Enum
 from ipaddress import IPv4Address, IPv4Interface, IPv4Network, IPv6Address, IPv6Interface, IPv6Network
+from pathlib import Path
 from uuid import UUID
 
 import pytest
 
 from pydantic import BaseModel, create_model
 from pydantic.json import pydantic_encoder, timedelta_isoformat
+from pydantic.types import DirectoryPath, FilePath, SecretBytes, SecretStr
 
 
 class MyEnum(Enum):
@@ -21,6 +23,10 @@ class MyEnum(Enum):
     [
         (UUID('ebcdab58-6eb8-46fb-a190-d07a33e9eac8'), '"ebcdab58-6eb8-46fb-a190-d07a33e9eac8"'),
         (IPv4Address('192.168.0.1'), '"192.168.0.1"'),
+        (SecretStr('abcd'), '"**********"'),
+        (SecretStr(''), '""'),
+        (SecretBytes(b'xyz'), '"**********"'),
+        (SecretBytes(b''), '""'),
         (IPv6Address('::1:0:1'), '"::1:0:1"'),
         (IPv4Interface('192.168.0.0/24'), '"192.168.0.0/24"'),
         (IPv6Interface('2001:db00::/120'), '"2001:db00::/120"'),
@@ -42,6 +48,22 @@ class MyEnum(Enum):
 )
 def test_encoding(input, output):
     assert output == json.dumps(input, default=pydantic_encoder)
+
+
+def test_path_encoding(tmpdir):
+    class PathModel(BaseModel):
+        path: Path
+        file_path: FilePath
+        dir_path: DirectoryPath
+
+    tmpdir = Path(tmpdir)
+    file_path = tmpdir / 'bar'
+    file_path.touch()
+    dir_path = tmpdir / 'baz'
+    dir_path.mkdir()
+    model = PathModel(path=Path('/path/test/example/'), file_path=file_path, dir_path=dir_path)
+    expected = '{{"path": "/path/test/example", "file_path": "{}", "dir_path": "{}"}}'.format(file_path, dir_path)
+    assert json.dumps(model, default=pydantic_encoder) == expected
 
 
 def test_model_encoding():